project_tz/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-seed==0.3.1
django-seeder==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
Faker==20.1.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
packaging==23.2
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.8.0
uritemplate==4.1.1
